requirements.txt
# GenAI Study Assistant - Level 3
# Multi-Modal AI Quiz Generation System

# Web Framework
flask>=3.0.0
werkzeug>=3.0.0

# AI/ML Models
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.24.0
sentence-transformers>=2.2.2

# Local Whisper (FREE - no API needed)
openai-whisper>=20231117

# RAG & Embeddings
einops>=0.7.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
python-pptx>=0.6.23

# YouTube Processing
youtube-transcript-api>=0.6.0
yt-dlp>=2023.10.0

# Utilities
numpy>=1.24.0
tqdm>=4.66.0
Pillow>=10.0.0

# Note: ffmpeg must be installed separately
# Mac: brew install ffmpeg
# Ubuntu: sudo apt install ffmpeg
